Reduce amount of SSR work for metros and counties
