+++ lib/common_elements.js
@@ -34,20 +34,20 @@ export function Nav (props) {
     }
   }
   return (
-    <div className='flex bg-purple-200'>
+    <div className='flex bg-purple-200' prefetch>
       <Link href='/'>
         <a className={classes[0]}>State Comparisons</a>
       </Link>
-      <Link href='/states/Alabama'>
+      <Link href='/states/Alabama' prefetch>
         <a className={classes[1]}>States</a>
       </Link>
-      <Link href='/metros/Bakersfield, CA'>
+      <Link href='/metros/Bakersfield, CA' prefetch>
         <a className={classes[2]}>Metros</a>
       </Link>
-      <Link href='/counties/TX/Harris County'>
+      <Link href='/counties/TX/Harris County' prefetch>
         <a className={classes[3]}>Counties</a>
       </Link>
-      <Link href='/places/CA/Los Angeles'>
+      <Link href='/places/CA/Los Angeles' prefetch>
         <a className={classes[4]}>Cities</a>
       </Link>
       <Link href='/data-sources'>
